refactor(devframe): namespace adapters and invert client dep from kit
Move the six adapters (cli/build/spa/vite/kit/embedded) into
`devframe/adapters/*` so they're clearly grouped and no longer share the
top-level subpath namespace. Relocate the framework-neutral RPC client
surface from `@vitejs/devtools-kit/client` into `devframe/client`,
including `rpc`, `rpc-shared-state`, `rpc-static`, `rpc-ws`, `static-rpc`,
`docks`, and `context`. Kit now re-exports `devframe/client` plus its
kit-only `remote` + `client-script` helpers, establishing that devframe
must never depend on vite or any `@vitejs/*` package. Add a Dep Boundary
section to AGENTS.md documenting the rule.
