+++ alias.ts
@@ -21,13 +21,13 @@ export const alias = {
   'devframe/utils/shared-state': r('devframe/src/utils/shared-state.ts'),
   'devframe/utils/state': r('devframe/src/utils/state.ts'),
   'devframe/utils/when': r('devframe/src/utils/when.ts'),
-  'devframe/cli': r('devframe/src/cli.ts'),
-  'devframe/build': r('devframe/src/build.ts'),
-  'devframe/spa': r('devframe/src/spa.ts'),
-  'devframe/vite': r('devframe/src/vite.ts'),
-  'devframe/kit': r('devframe/src/kit.ts'),
-  'devframe/embedded': r('devframe/src/embedded.ts'),
-  'devframe/client': r('devframe/src/client.ts'),
+  'devframe/adapters/cli': r('devframe/src/adapters/cli.ts'),
+  'devframe/adapters/build': r('devframe/src/adapters/build.ts'),
+  'devframe/adapters/spa': r('devframe/src/adapters/spa.ts'),
+  'devframe/adapters/vite': r('devframe/src/adapters/vite.ts'),
+  'devframe/adapters/kit': r('devframe/src/adapters/kit.ts'),
+  'devframe/adapters/embedded': r('devframe/src/adapters/embedded.ts'),
+  'devframe/client': r('devframe/src/client/index.ts'),
   'devframe': r('devframe/src'),
   '@vitejs/devtools-kit/node': r('kit/src/node/index.ts'),
   '@vitejs/devtools-kit/client': r('kit/src/client/index.ts'),
