+++ packages/devframe/src/adapters/cli.ts
@@ -1,4 +1,4 @@
-import type { DevtoolDefinition } from './types/devtool'
+import type { DevtoolDefinition } from '../types/devtool'
 import process from 'node:process'
 import c from 'ansis'
 import cac from 'cac'
@@ -6,11 +6,11 @@ import { getPort } from 'get-port-please'
 import { createApp } from 'h3'
 import { resolve } from 'pathe'
 import sirv from 'sirv'
+import { DEVTOOLS_MOUNT_PATH } from '../constants'
+import { createHostContext } from '../node/context'
+import { createH3DevToolsHost } from '../node/host-h3'
+import { startHttpAndWs } from '../node/server'
 import { buildStatic } from './build'
-import { DEVTOOLS_MOUNT_PATH } from './constants'
-import { createHostContext } from './node/context'
-import { createH3DevToolsHost } from './node/host-h3'
-import { startHttpAndWs } from './node/server'
 import { buildSpa } from './spa'
 
 export interface CreateCliOptions {
