+++ packages/devframe/src/client/index.ts
@@ -1,0 +1,4 @@
+export * from './context'
+export * from './docks'
+export * from './rpc'
+export { getDevToolsRpcClient as connectDevtool } from './rpc'
