+++ packages/devframe/src/client/static-rpc.ts
