+++ packages/devframe/tsdown.config.ts
@@ -18,13 +18,13 @@ export default defineConfig({
     'utils/shared-state': 'src/utils/shared-state.ts',
     'utils/state': 'src/utils/state.ts',
     'utils/when': 'src/utils/when.ts',
-    'cli': 'src/cli.ts',
-    'build': 'src/build.ts',
-    'spa': 'src/spa.ts',
-    'vite': 'src/vite.ts',
-    'kit': 'src/kit.ts',
-    'embedded': 'src/embedded.ts',
-    'client': 'src/client.ts',
+    'adapters/cli': 'src/adapters/cli.ts',
+    'adapters/build': 'src/adapters/build.ts',
+    'adapters/spa': 'src/adapters/spa.ts',
+    'adapters/vite': 'src/adapters/vite.ts',
+    'adapters/kit': 'src/adapters/kit.ts',
+    'adapters/embedded': 'src/adapters/embedded.ts',
+    'client/index': 'src/client/index.ts',
   },
   tsconfig: '../../tsconfig.base.json',
   clean: true,
