+++ packages/kit/src/client/index.ts
@@ -1,5 +1,3 @@
 export * from './client-script'
-export * from './context'
-export * from './docks'
 export * from './remote'
-export * from './rpc'
+export * from 'devframe/client'
